chore: move wallets/react logics into 2 separate hooks
