+++ wallets/react/src/legacy/types.ts
@@ -33,6 +33,13 @@ export type ProviderContext = {
   disconnectAll(): Promise<PromiseSettledResult<any>[]>;
   state(type: WalletType): WalletState;
   canSwitchNetworkTo(type: WalletType, network: Network): boolean;
+  /**
+   * `Provider` in legacy terms means injected instances by wallets into window (e.g. window.ethereum)
+   * that can be retrieved by `getInstance`.
+   *
+   * Note 1: Providers are lazy evaluated, which means you need to call `connect` (or `state`) first, then the value will be shown in object.
+   *         before doing that, it's a key (wallet name or we call it `type` to be more specific) with null value. (e.g. {metamask: null})
+   */
   providers(): Providers;
   getSigners(type: WalletType): SignerFactory;
   getWalletInfo(type: WalletType): WalletInfo;
